+++ inferno-ml-server/src/Inferno/ML/Server/Bridge.hs
@@ -7,7 +7,6 @@ module Inferno.ML.Server.Bridge
 where
 
 import Control.DeepSeq (NFData)
-import Control.Monad.Catch (throwM)
 import Control.Monad.IO.Class (liftIO)
 import Control.Monad.Reader (asks)
 import Data.Int (Int64)
@@ -16,7 +15,7 @@ import Database.PostgreSQL.Simple.SqlQQ (sql)
 import Inferno.Core (Interpreter, mkInferno)
 import qualified Inferno.ML.Server.Client.Bridge as Bridge
 import Inferno.ML.Server.Module.Bridge (mkBridgeFuns)
-import Inferno.ML.Server.Module.Prelude (mkBridgePrelude)
+import Inferno.ML.Server.Module.Prelude (mkServerBridgePrelude, serverMlPrelude)
 import Inferno.ML.Server.Types
 import Inferno.ML.Server.Utils
 import Inferno.ML.Types.Value (customTypes)
@@ -37,12 +36,12 @@ import System.Posix.Types (EpochTime)
 initializeInferno ::
   Id InferenceParam -> RemoteM (Interpreter RemoteM BridgeMlValue)
 initializeInferno ipid = do
-  (`mkInferno` customTypes) . mkBridgePrelude . mkFuns
+  (`mkInferno` customTypes) . (`mkServerBridgePrelude` serverMlPrelude) . mkFuns
     =<< getBridgeInfo
   where
     -- There should always be a bridge saved for the param
     getBridgeInfo :: RemoteM BridgeInfo
-    getBridgeInfo = firstOrThrow NoBridgeSaved =<< queryStore q (Only ipid)
+    getBridgeInfo = firstOrThrow (NoBridgeSaved ipid) =<< queryStore q (Only ipid)
       where
         q :: Query
         q = [sql| SELECT * FROM bridges WHERE id = ? |]
@@ -51,20 +50,20 @@ initializeInferno ipid = do
     mkFuns bi = mkBridgeFuns valueAt latestValueAndTimeBefore valuesBetween
       where
         valueAt :: Int64 -> PID -> EpochTime -> RemoteM IValue
-        valueAt res pid = callBridge bi . Bridge.valueAtC res pid
+        valueAt res pid = callBridge ipid bi . Bridge.valueAtC res pid
 
         latestValueAndTimeBefore :: EpochTime -> PID -> RemoteM IValue
         latestValueAndTimeBefore t =
-          callBridge bi . Bridge.latestValueAndTimeBeforeC t
+          callBridge ipid bi . Bridge.latestValueAndTimeBeforeC t
 
         valuesBetween :: Int64 -> PID -> EpochTime -> EpochTime -> RemoteM IValue
         valuesBetween res pid t1 =
-          callBridge bi . Bridge.valuesBetweenC res pid t1
+          callBridge ipid bi . Bridge.valuesBetweenC res pid t1
 
 -- | Call one of the bridge endpoints using the given 'BridgeInfo'
-callBridge :: (NFData a) => BridgeInfo -> ClientM a -> RemoteM a
-callBridge bi c =
-  either (throwM . ClientError) pure =<< liftIO . runClientM c =<< mkEnv
+callBridge :: (NFData a) => Id InferenceParam -> BridgeInfo -> ClientM a -> RemoteM a
+callBridge ipid bi c =
+  either (throwRemoteError . ClientError ipid . show) pure =<< liftIO . runClientM c =<< mkEnv
   where
     mkEnv :: RemoteM ClientEnv
     mkEnv = asks $ (`mkClientEnv` url) . view #manager