+++ inferno-ml-server/src/Inferno/ML/Server/Utils.hs
@@ -1,7 +1,11 @@
+{-# LANGUAGE ConstraintKinds #-}
+{-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
 module Inferno.ML.Server.Utils
-  ( throwInfernoError,
+  ( HasPool,
+    throwInfernoError,
+    throwRemoteError,
     firstOrThrow,
     queryStore,
     executeStore,
@@ -12,8 +16,10 @@ where
 import Control.Monad (void)
 import Control.Monad.Catch (Exception, MonadThrow (throwM))
 import Control.Monad.IO.Class (liftIO)
+import Control.Monad.Reader (MonadReader)
 import Data.Generics.Labels ()
-import Data.Pool (withResource)
+import Data.Generics.Product (HasType (typed))
+import Data.Pool (Pool, withResource)
 import Data.Vector (Vector, (!?))
 import Database.PostgreSQL.Simple
   ( Connection,
@@ -28,20 +34,29 @@ import Inferno.ML.Server.Types
 import Lens.Micro.Platform (view)
 import UnliftIO (MonadUnliftIO (withRunInIO))
 
-throwInfernoError :: forall e a. (Exception e) => Either e a -> RemoteM a
-throwInfernoError = either (throwM . InfernoError . SomeInfernoError) pure
+throwRemoteError :: RemoteError -> RemoteM a
+throwRemoteError = throwM
 
-queryStore :: (ToRow b, FromRow a) => Query -> b -> RemoteM (Vector a)
+throwInfernoError :: (Exception e) => Id InferenceParam -> e -> RemoteM a
+throwInfernoError ipid = throwRemoteError . InfernoError ipid . SomeInfernoError . show
+
+type HasPool r m = (MonadUnliftIO m, MonadReader r m, HasType (Pool Connection) r)
+
+queryStore :: forall b a m r. (HasPool r m, ToRow b, FromRow a) => Query -> b -> m (Vector a)
 queryStore q x = withConns $ \conn -> liftIO $ query conn q x
+{-# INLINE queryStore #-}
 
-executeStore :: (ToRow a) => Query -> a -> RemoteM ()
+executeStore :: forall a m r. (HasPool r m, ToRow a) => Query -> a -> m ()
 executeStore q x =
   withConns $ \conn ->
     liftIO . withTransaction conn . void $
       execute conn q x
+{-# INLINE executeStore #-}
 
-firstOrThrow :: (MonadThrow m, Exception e) => e -> Vector a -> m a
+firstOrThrow :: (MonadThrow m) => RemoteError -> Vector a -> m a
 firstOrThrow e = maybe (throwM e) pure . (!? 0)
+{-# INLINE firstOrThrow #-}
 
-withConns :: (Connection -> RemoteM b) -> RemoteM b
-withConns f = view #store >>= \cs -> withRunInIO $ \r -> withResource cs $ r . f
+withConns :: (HasPool r m) => (Connection -> m b) -> m b
+withConns f = view typed >>= \cs -> withRunInIO $ \r -> withResource cs $ r . f
+{-# INLINE withConns #-}