+++ nix/inferno-ml/images/common/cpu.nix
@@ -7,6 +7,7 @@
     enable = true;
     configuration = {
       port = 8080;
+      instanceId = "auto";
       cache = {
         path = "/home/inferno/.cache/models";
         max-size = 10 * 1073741824;