+++ inferno-ml-server/CHANGELOG.md
@@ -1,7 +1,19 @@
 # Revision History for `inferno-ml-server`
 
-## 2025.2.28
+## 2025.3.28
 * Update for new order of `makeWrites` elements
+
+## 2025.3.11
+* Persist error traces in DB if instance-id is provided in the config
+
+## 2025.3.6
+* Add `toDevice` impl with logging on failure to move tensor
+
+## 2025.3.3
+* Brought back shims from inferno-ml-server-types
+
+## 2025.2.28
+* Extracted RemoteTrace and deps to inferno-ml-server-types
 
 ## 2025.2.13
 * Fourmolu